1. 客户管理增加 折扣字段，浮点型，2位小数 2. 基础数据增加佣金管理页面 3. 财务管理增加数据统计页面

+++ VOS.ViewModel/Business/VOS_TaskVMs/VOS_TaskListVM.cs
@@ -230,7 +230,7 @@ namespace VOS.ViewModel.Business.VOS_TaskVMs
             else
             {
                 return data.OrderBy(x => ((int)x.OrderState == 1 ? 1 :
-                                      ((int)x.OrderState == 0 ? 2 :
+                                      (x.OrderState == 0 ? 2 :
                                       ((int)x.OrderState == 2 ? 3 :
                                       (int)x.OrderState == 3 ? 4 : 5))));
             }
